fixed calibration

+++ calibration/src/Calibrator.h
@@ -32,7 +32,7 @@ private:
 
     PointCloud<POINTTYPE>::Ptr cropBox(PointCloud<POINTTYPE>::Ptr pc);
     pcl::ModelCoefficients::Ptr findGroundPlane(PointCloud<POINTTYPE>::Ptr pc);
-    std::vector<double> computeAngles(std::vector<double> groundPlaneCoefficients);
+    std::vector<double> computeAngles(std::vector<float> groundPlaneCoefficients);
     void saveCalibration(std::vector<double> angles, double height);
 
 };
